refactor: step2 리뷰 반영; 패키지 명 변경
