--- src/main/java/lotto/domain/NumberGenerationStrategy.java
@@ -1,10 +1,0 @@
-package lotto.domain;
-
-import java.util.List;
-
-import lotto.domain.vo.LottoNumber;
-
-@FunctionalInterface
-public interface NumberGenerationStrategy {
-    List<LottoNumber> getNumbers();
-}
