--- src/main/java/lotto/domain/vo/LottoNumber.java
@@ -1,45 +1,0 @@
-package lotto.domain.vo;
-
-import java.util.Objects;
-
-public class LottoNumber implements Comparable<LottoNumber> {
-    private static final int MIN_LOTTO_NUMBER = 1;
-    private static final int MAX_LOTTO_NUMBER = 45;
-
-    private final int number;
-
-    public LottoNumber(int number) {
-        validate(number);
-        this.number = number;
-    }
-
-    private void validate(int number) {
-        if (number < MIN_LOTTO_NUMBER || number > MAX_LOTTO_NUMBER) {
-            throw new IllegalArgumentException("로또 번호 범위 밖의 숫자의 입력입니다.");
-        }
-    }
-
-    public int getValue() {
-        return number;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-        LottoNumber that = (LottoNumber)o;
-        return number == that.number;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(number);
-    }
-
-    @Override
-    public int compareTo(LottoNumber o) {
-        return Integer.compare(this.number, o.number);
-    }
-}
