--- src/main/java/lotto/view/exception/InvalidInputException.java
@@ -1,7 +1,0 @@
-package lotto.view.exception;
-
-public class InvalidInputException extends RuntimeException {
-    public InvalidInputException(String message) {
-        super(message);
-    }
-}
