+++ src/main/java/lottogame/LottoGameApplication.java
@@ -1,0 +1,11 @@
+package lottogame;
+
+import lottogame.controller.LottoGameController;
+import lottogame.view.InputView;
+import lottogame.view.OutputView;
+
+public class LottoGameApplication {
+    public static void main(String[] args) {
+        new LottoGameController(new InputView(), new OutputView()).run();
+    }
+}
