+++ src/main/java/lottogame/domain/NumberGenerationStrategy.java
@@ -1,0 +1,10 @@
+package lottogame.domain;
+
+import java.util.List;
+
+import lottogame.domain.lotto.LottoNumber;
+
+@FunctionalInterface
+public interface NumberGenerationStrategy {
+    List<LottoNumber> getNumbers();
+}
