+++ src/main/java/lottogame/view/exception/InvalidInputException.java
@@ -1,0 +1,7 @@
+package lottogame.view.exception;
+
+public class InvalidInputException extends RuntimeException {
+    public InvalidInputException(String message) {
+        super(message);
+    }
+}
